Added purchase post feature

+++ src/main/java/br/com/letscode/dataproject/product/model/Product.java
@@ -22,7 +22,7 @@ public class Product {
     private String codeNumber;
     
     @NonNull
-    private BigDecimal price;
+    private Float price;
     
     @NonNull
     private Integer quantity;
